+++ reconcile/test/rhidp/test_common.py
@@ -4,6 +4,13 @@ from unittest.mock import Mock
 import pytest
 from pytest_mock import MockerFixture
 
+from reconcile.gql_definitions.fragments.ocm_environment import OCMEnvironment
+from reconcile.gql_definitions.fragments.vault_secret import VaultSecret
+from reconcile.gql_definitions.rhidp.clusters import (
+    ClusterAuthOIDCV1,
+    ClusterV1,
+    OpenShiftClusterManagerV1,
+)
 from reconcile.rhidp import common
 from reconcile.test.ocm.fixtures import build_cluster_details
 from reconcile.utils.ocm.labels import (
@@ -84,3 +91,50 @@ def test_rhidp_common_discover_clusters_without_org_filter(
     clusters = common.discover_clusters(None, None)  # type: ignore
 
     assert org_id in clusters
+
+
+def test_rhidp_common_build_cluster_obj(
+    ocm_env: OCMEnvironment, build_cluster_rhidp_labels: LabelContainer
+) -> None:
+    cluster_details = build_cluster_details(
+        cluster_name="cluster_name",
+        subscription_labels=build_cluster_rhidp_labels,
+        org_id="org_id",
+    )
+    assert common.build_cluster_obj(
+        ocm_env,
+        cluster_details,
+        auth_name="auth_name",
+        auth_issuer_url="https://foobar.com",
+    ) == ClusterV1(
+        name="cluster_name",
+        consoleUrl="https://console.foobar.com",
+        ocm=OpenShiftClusterManagerV1(
+            name="",
+            environment=OCMEnvironment(
+                name="env",
+                url="https://ocm",
+                accessTokenClientId="client-id",
+                accessTokenUrl="https://sso/token",
+                accessTokenClientSecret=VaultSecret(
+                    path="path", field="client-secret", version=None, format=None
+                ),
+            ),
+            orgId="org_id",
+            accessTokenClientId=None,
+            accessTokenUrl=None,
+            accessTokenClientSecret=None,
+            blockedVersions=None,
+            sectors=None,
+        ),
+        upgradePolicy=None,
+        disable=None,
+        auth=[
+            ClusterAuthOIDCV1(
+                service="oidc",
+                name="auth_name",
+                issuer="https://foobar.com",
+                claims=None,
+            )
+        ],
+    )
