+++ reconcile/test/rhidp/test_ocm_oidc_idp_integration.py
@@ -17,11 +17,14 @@ from reconcile.rhidp.ocm_oidc_idp.integration import (
 
 def test_ocm_oidc_idp_integration_get_clusters(cluster_query_func: Callable) -> None:
     intg = OCMOidcIdpIntegration(
-        OCMOidcIdpIntegrationParams(vault_input_path="foo/bar")
+        OCMOidcIdpIntegrationParams(
+            vault_input_path="foo/bar", default_auth_issuer_url="https://issuer.com"
+        )
     )
     assert intg.get_clusters(cluster_query_func) == [
         ClusterV1(
             name="cluster-1",
+            consoleUrl="https://console.url.com",
             ocm=OpenShiftClusterManagerV1(
                 name="ocm-production",
                 accessTokenClientId=None,
@@ -61,6 +64,7 @@ def test_ocm_oidc_idp_integration_get_clusters(cluster_query_func: Callable) -> 
         ),
         ClusterV1(
             name="cluster-2",
+            consoleUrl="https://console.url.com",
             ocm=OpenShiftClusterManagerV1(
                 name="ocm-production",
                 accessTokenClientId=None,
@@ -98,37 +102,5 @@ def test_ocm_oidc_idp_integration_get_clusters(cluster_query_func: Callable) -> 
                     ),
                 ),
             ],
-        ),
-        ClusterV1(
-            name="cluster-3",
-            ocm=OpenShiftClusterManagerV1(
-                name="ocm-production",
-                environment=OCMEnvironment(
-                    name="name",
-                    url="https://api.openshift.com",
-                    accessTokenClientId="access-token-client-id",
-                    accessTokenUrl="http://token-url.com",
-                    accessTokenClientSecret=VaultSecret(
-                        field="client_secret",
-                        format=None,
-                        path="path/to/client_secret",
-                        version=None,
-                    ),
-                ),
-                accessTokenClientId="access-token-client-id",
-                accessTokenUrl="http://token-url.com",
-                accessTokenClientSecret=VaultSecret(
-                    field="client_secret",
-                    format=None,
-                    path="path/to/client_secret",
-                    version=None,
-                ),
-                orgId="org-id",
-                blockedVersions=[],
-                sectors=None,
-            ),
-            upgradePolicy=None,
-            disable=None,
-            auth=[ClusterAuthV1(service="github-org-team")],
         ),
     ]
